Replace pangram with a stanza from the Jabberwocky
This isn't a proper pangram (it's missing 4 letters), but it's more fun than a regular
pangram. I may shave a yak later and see about writing a Python script that can search
text for a pangram and see if there are any in Alice in Wonderland.

I've also removed the specimen that enumerated letters, numbers, and symbols.