Add a bottom margin to headings
Headings should stand out on the page with plenty of whitespace around them, so I'm
breaking with tradition and give the heading a top and bottom margin. I had to remove top
margins from anything that follows a heading because these are often in grid layouts, so
the margins don't collapse.