Only apply padding to headings as needed
Use a media query to apply the padding at the same time we move the header anchor to the
right of the heading. This way we can let the headings stretch out, if needed. I also
moved the CSS for the headings related to positioning the header anchor to the header
anchor SCSS file, rather than leaving it in base. Realistically this might call for a
component, but honestly this site is so small it doesn't seem worthwhile to muck about in
markdown-it to apply classes to all my headings.