Fix the alignment in the footer
Good news: grid does work the way I thought, I was being silly and changing the font-size
on the <footer> element, which made the max size of the content column (60ch) smaller than
the others. By setting the font-size on all the children of the footer instead of the
footer itself, the grid lines up with the other two grids on the page.