Create cards for each project
I've added a card class to the CSS for creating a clickable card. It makes no real
assumptions about the card itself, it just handles the CSS necessary to make the main
anchor the size of the card using the ::after trick. I also had to make some changes to
how we set up headings for the anchor so that only plain headings are position: relative,
otherwise the card trick doesn't work.