Break up the card gradients into custom properties
Use three custom properties to set the background gradients on cards. I've found that
using properties like this makes it easier to apply the cicada principle and results in
a more random appearance. I've also abandoned the random re-sizing in favor of just
allowing the cards to be sized based on their title lengths.