Add some conditionals to the post template
Conditionally show the publish date if there is one. Wrap the footer in a block so that it
can be overridden. This way I can reduce some of the duplication between the project page
template and the blog post template. I moved the content for the blog post footer into the
weblog data file.