Center the body with a max width
Give the body a max width based on the measure. It's larger than the measure to allow
components like the project card grid to have a little more space, and give the pages a
jaunty, asymmetric feel on wide viewports, since the prose will appear slightly
off-center.

I created a variable in the tokens to reference the site margin, since the header will
need that to set the proprotions on the grid.