Increase the breakpoint for the header anchors
Without the base grid layout, we need to wait longer before we can shift the header
anchors over to the left side of the header. One difference between this layout and the
grid-based one is that the left edge of the content reaches the left edge of the viewport
much quicker.