Update the type scale
Making a few adjustments to the type scale, but I'm thinking I may abandon the global use
of Utopia for all type. I think what I'd like is a single scale for most thigns, but <h1>
can have a scale that increases with the viewport so that the page titles become more
dramatic with larger viewports. The rest of the fonts can have a more subtle font scale
maybe.