Tighten up spacing on lists
I've reduced the indent on lists, because the default indentation is too much, in my
opinion. I've tightened up the leading on each list item, and given a little more space
between list items so it's more clear which lines are grouped in a single item. Lists also
get more margin above and below them so that they're more easily distinguished from the
surrounding paragraphs.