Adjust styles on blog index
The new utility classes have made it easy for me to get some useful styles going so that
the summary for each post is grouped nicely in a visual way without resorting to creating
some new component. Just a smattering of utility classes to adjust line heights and
margins, and we've got some nice groupings.