Add the second project layout: dorian
I've moved most of the configuration for ionian into the base project-summary class. This
gives a somewhat more consistent layout to all projects in the portfolio page — at least
at smaller viewport sizes. The dorian layout let's the image get wider and keeps the text
below it like it does on the smaller viewports. I've made the body a multi-col layout now,
so for the dorian layout the project description typically gets split across multiple
columns.

I might want to beef up the description for Polio just to take better advantage of this.