Slim down the reset
I'm finding that there are fewer things I really consider a reset, and more things that I
just want to be the base styles for a given site. No point in having long, specific lists
of elements that have their bottom margin removed only to be overridden later with
something that has greater specificity.

I am also, for now, considering the default top margin to be a reset, although I plan to
remove this as it is requiring too many overrides.