Remove the h-card from posts
It seems like I probably don't need an h-card in the h-entry? Keith J. Grant doesn't have
one on his posts, and he's the one whose guide I'm following, so probably if there isn't
an author for the post, it falls back to the site's author? I guess we'll find out.