Move "global" partials up one level
No real need to distinguish between global partials and components, if
I'm sticking with the idea that all partials can be dropped in anywhere
without the need to set up context. Anything that is parameterized (like
a component) should be a macro to make it easier to pass in objects.