Inline the webmention template in the list template
It makes the webmentions partial a little messier to read, but at the
moment I have no reason to believe I'll ever need the markup for an
individual webmention outside of that big block of webmentions. So
rather than move it into a macro on its own, I might as well just reduce
the total number of templates by one.