Layout navigation in the header and footer
Apply a cluster layout from Every Layout for any <ul> in a <nav> element and
remove the list styles. It's possible that I won't want this to always be true
for navs, if I end up implementing a table of contents. We'll see.

I also started converting some values to variables to try making it easier to
update to something using clamp() in the future.