Replace name mark with an image of me
Taking some inspiration from Thomas Semmler's site (helloyes.dev) and
using an image of myself for the home link, rather than the name mark. I
love that name mark, but I think I'm dropping Sporting Grotesque (:cry:)
and it probably makes more sense to create something like that as an SVG
and drop it in, rather than including a bunch of CSS just to create that
effect. I think the changing polygon behind it was too subtle for anyone
to notice.