Back to grayscale for now
Rather than spending lots of time tweaking the colors, I'm going to go back to
basics and design in grayscale. Worst case, the site is very Hedon-esque, which
is pretty great, to be honest. I already kind of love the aesthetic, although
I'd still like to get some color in there.