Add a shortcode for clamp expressions
This shortcode does the math necessary to create a responsive size based
on viewport units, rems, and a min and max viewport width. Doing the
calculations in a shortcode will allow me to store tokens as a range
instead of as a clamp expression in a string, which will be more
flexible. It also makes it a lot easier to change my scales if I want
to, since I won't have to rebuild each clamp expression by hand.