Swap Drafting for League Mono
I think Drafting has too much character for my purposes. I'm using a monospace
font for code and metadata, and I think Drafting works better as a main font.
The letterforms have a lot of nice touches that are probably great for reading,
but distracting when they represent secondary information.