Refactor typography styles
I refactored my typography in the CSS so that it will (hopefully) be
easier to maintain. I'm building up a collection of presets à la Dan
Mall's approach[^1] that I've labelled with comments so that I can just
add selectors to those blocks as needed. This should help me keep a lid
on how many different styles I use in my typography because I won't be
setting type styles anywhere else.

[^1]: https://superfriendlydesign.systems/articles/typography-in-design-systems/)