Style the post summaries
Apply some classes and layout to style the post summaries in the list of posts.
There are attributes now for easily changing the spacing in a flow layout so
that I can adjust the spacing of the elements in the post summary. I also
applied the display-2 typography to the headings so that they get Sporting
Grotesque; I had to decrease the font size on that heading because Sporting
Grotesque is such a heavy typeface.