Create full bleed figures
Switch the grid technique to a constrained grid I learned from Ethan Marcotte
that makes it possible to nest the grids and keep the columns aligned and then I
can use the grid columns to stretch out the figure. I added a little background
color to offset the figure sfrom the surrounding text.