Remove the colophon from the footer
I'm going to put the colophon somewhere in the vicinity of /about/, so I no
longer need it in the footer of every page. I made some adjustments to the
footer layout to place the social icons to the right of the copyright, now that
I no longer have the colophon.