Display the permalink after the heading
This is a little nicer for screen readers, since the permalink to the heading
won't be announced as part of the heading, but is announced afterwards. I'm
compromising a little by displaying the word "Permalink", when what I'd like to
show is the ID of each heading, but that's going to require a custom permalink
renderer, and I don't want to bother with that right now.

I'm still tweaking the layout. Unfortunately, I think, in order to get the
headings, the permalink, and the paragraphs aligned the way I want, I might need
to wrap the heading and permalink together in a <div> so that the end up in the
same grid cell.