Switch to a headling link for permalinks
As pointed in the markdown-it-anchor docs, this is the style used by MDN and the
HTTP Archive, so maybe this is a good approach. It certainly saves me the
trouble of implementing a custom permalink function to get the markup I want for
my multi-column layout. This causes problems with Safari's reader mode, which I
do care about, but maybe not enough to skip the solution since it's so easy and
this is arguably Safari's problem, not an inappropriate use of HTML.