Add space tokens
I've added two types of space tokens: fluid and relative. Fluid spaces scale
with the viewport size but are based on rems, so no matter what other CSS is set
on an element using one of these tokens, the spacing will remain the same.
Relative spaces use ems, so they scale with the font size of the element they're
applied to, which I think will be useful for setting margins on prose.

I made some changes to the schema of the tokens so that both the fluid spaces
and the fluid font sizes are defined in the same property so I don't have to
duplicate values like the viewport range.