Rearrange tokens to allow easier iteration
I've created two top-level containers for my tokens to separate fluid scales
from static tokens, this way I can iterate over each of those properties to
write out the props and whenever I add new tokens, they'll get picked up
automatically. I changed the way the fluid spaces are generated so that they fit
the same pattern as the font scales and added the ability to name levels to
accommodate spaces.

The naming of levels in fluid scales is a bit flimsy as it assumes the starting
step is 0. Since I'm not building a general framework for fluid scales in Sass,
I think this is probably fine. At least for now.