Fix spacing on the navigation links
Reset margins on list items in cluster layouts, again because I often
apply cluster to a list. Fix application of the gap property so that
it uses --space and we can take advantage of our space utilities to
adjust the gap per-component and add utilities for fluid spaces as well
as relative spaces.