Fix iframe sizing
For now at least this is probably the best I can do. There may be
problems for mobile users in landscape mode with the embedded iframes.
This may require some tweaking, or I may end up preferring to implement
my own sketchbook so that it's easier to integrate into my site.