Style site header
Restore some of the styles for the site header. The cluster layout is back in
place and I've sized the image of me in the top left and restored the hover
animation. I hid the skip link as well. The skip link styles still have some CSS
custom props that I haven't replaced, I left them in as place holders to remind
myself to replace them.