Refactor the gallery component into a layout template
I think I can create a layout template that can handle the gallery layouts for
both the sketchbook and the home page latest posts list, so there's no reason to
make a macro. It's a little tempting to create a shortcode for generating the
cards, because that would be easy to write some tests for and perhaps make the
template a little more readable.