Subset Sporting Grotesque and Epilogue
I added a script that uses pyftsubset and glyphhanger to subset my main fonts,
removing opentype features I don't want and eliminating characters I shouldn't
need. Epilogue is about 1/3 as big as it was. Savings on Sporting Grotesque are
less significant, but it's still good to eliminate some dead weight.

I might be able to trim the character set for Sporting Grotesque even further
because I mainly use it for titles.