Add scaled down images for the header
No need to ship larger images for the header, although it might be good to
figure out a way to implement this with a <picture> element or srcsets so that
we get a larger image if someone zooms the screen or increases their font size.
I also set a width and height on the image so that we avoid layout shifts and
getting yelled at by lighthouse.