Rename image shortcode to picture
I kind of like the way Astro has separate `<Image />` and `<Picture />`
components for generating either a single responsive `<img>` tag or a
`<picture>` tag with multiple sources, so I'm copying that here. I don't have an
image tag, but I may want one some day. And if not, it's not doing any harm.