Convert index pages to JS templates
I decided to use a layout for index pages that expects a `pageList` variable in
its context. The easiest way, then, to set that variable is in eleventyComputed,
and since these pages don't need any markdown on them, it's easiest to just make
them all JS templates. I eliminated the feed component, now that it's not in
use.