Add an extra block class for posts
I added the post-block class to be able to style post lists a little. I'm mainly
just pulling the pub date up to the top (yes, yes, I know I shouldn't reorder
things; it seems like such a small thing that it's unlikely to bother people too
much).

I was thinking about setting the post titles in Sporting Grotesque, but it
conflicts a bit too much with the page titles, so I opted to leave it as
Epilogue.