Eliminate the page layout
Turned the base layout into something more than just the HTML boilerplate so
that it could be used by pages, eliminating the need for the page layout. The
article layout is used for anything with a publication date that should display
web mentions, eliminating the need for the showWebmentions flag.