Move the list reset into the reset styles
Rather than a utility class, it makes sense to apply these styles whenever we
explicitly set aria-role=list, because we only do this when we're removing list
styles so that Safari still register the list as a list. If we don't need it to
register as a list, we don't need a ul/ol/dl element, and thus we don't need the
reset styles.