Fix portfolio layout
I added the portfolio template as a layout, rather than generating the portfolio
in the page, at least for now. This was the easiest way to break out of the
center layout in the base layout, although maybe this is a sign that my base
layout is a little too prescriptive.

I updated the CSS variables used in the project layouts so that everything is
laid out correctly.