Convert tokens to variable/mixins-only module
The tokens module doesn't provide any css modules — or at least very little. The
individual modules no longer output their own properties, instead I'm using
Miriam Suzanne's trick of iterating over all of the variables using
sass:meta.module-variables. Storing the tokens as variables instead of inside a
map makes them a little easier to work with, although I'm not sure I'll be
working with them directly. I've also renamed them all so that the names just
become the custom prop names.

Now there's only one loop in default/_index.scss that produces all of the tokens
as props on :root, which is nice and convenient. Adding new tokens is as simple
as adding new variables to a module and making sure that module is forwarded
from the tokens module.

The only CSS still in the tokens module are my @font-face rules because I can't
think of a better place to put them.