Reintroduce the color scheme styles
Uncomment the scheme.css mixin. I'm tempted to put some layers in here to ensure
that the user overrides set with the data-theme attribute are guaranteed to
apply over the OS preferences (I hit a bug recently where they weren't, but I
can't remember what the issue was), but I guess for now I'll leave it and only
slap a layer in there if I need to.