Reintroduce the document base styles
Remove all of the text presets because that's what I need to work on next and
they're not working right now. This way I can hit reset on that and set up a
little design page with some of the tokens/patterns for the site and sand out
some of the rough edges.