Set up type examples and utility classes
I'm adding the utility classes back in because it will be the easiest way to
apply the type styles for the design page. I added another block to the base
layout so that I can avoid the site header and footer as well as the other page
paraphernalia.

Also added some inline styles to the design page to make it a little more
legible for now. It's layout is not likely to be the same as the rest of the
site, since it's not really a public page.