Reduce the number of steps in the font scale
I had more steps than I was using so I reduced the number of steps by 1 and
increased the scales to get larger type. This gives me a nice large display font
for posts along with a big heading size for regular <h1>s and some options for
other headings.