Add some margin to the design page
It was really hard to read the design page without some margins, so I added my
layout layer back in so that I could use the stack utility. I changed the
default value to be 1em, since that gives a better rhythm for differently sized
type.

I also swapped the order of the type style labels and their demos so that the
style come first. Maybe once I get some better utilities for the typography I
can create a better hierarchy here and be able to switch them back like Mike's
Frontend Horse example, but for now it's easier to tell what's what with them
re-ordered.