Adjust body type styles so we have 4
I had some weird gaps in the body type mixins because I'd deleted some things,
so I reshuffled all of my body styles so that they go from large (1) to small
(4).

body-2 is the second largest and in italics. The idea being that this might be
good for pull quotes, or maybe even block quotes. I'm not sure.