Remove the largest body style
I think having a body style at font-size-2 was too big, it overlapped a lot with
the headings and I don't know what use I have for it. I also extracted the
italic styles into a mixin, because we also want to use the character variants
in the italic (I like the one story "a" for italics).