Move site configuration out of _document.scss
I think it makes sense to separate the site configuration — the assignment of
tokens to semantic variables — to a separate file. The config custom props are
set on body, although maybe HTML makes more sense. I should do a little testing
to see if :root or html has more specificity, but I'm pretty sure it's :root,
which would mean I'd have to put the site config in a separate layer if I want
it on html instead of body.