Reset some basic font styles
I applied all of the text presets as-needed to remove all of the TODOs. I'm
setting some default top margins for a number of elements that are primarily
used in prose so that I don't have to always apply a .stack utility or anything.
It shouldn't be hard to reset these margins as-needed with utility classes and
on a site that is mainly prose, it makes sense to set default styles for prose.