Remove the list reset from the cluster layout
Everywhere that I used a list as a cluster layout should've had role=list, so I
updated all of those lists with the role. This means the list reset applied to
them, so I could remove the `list-style: none` and `padding: 0` from the
cluster. I did have to set `margin: 0` on all immediate children of a cluster,
since I'm taking the position that many things ought to have margins by default
and only removing them as necessary (which feels reasonable for a site that is
largely prose, not an application interface). In the cluster, spacing is handled
by the gap property, so removing all margins seems reasonable.