Restructure project directories
I kinda like the project structure that Astro uses, so I'm shoehorning 11ty to
fit. It keeps my page templates separate from the other paraphernalia of the
site, which is quite nice. Also, using a public/ directory for things that get
passed through makes it much easier to configure Eleventy; instead of having to
add a passthrough copy config for everything that needs to be copied, I just
dump it all in the public/ folder and configure a passthrough for that.