Move all styles into a single style.css file
Instead of loading each layer from a separate file using an inline style tag in
the site, I've gone back to compiling it all into a single file. I have a
suspicion that part of the reason the thumbnail didn't work on personalsit.es is
because of that silly approach.