Refactor the layouts so the page header is absent on home
I never really wanted the home page to display "Home" in large letters, that was
just an accident of poor template refactoring. Now I've fixed it a bit so that
the home page has no big header like that. I do think there is probably a bit of
an accessibility problem with no h1 on the page…