Move the layer definition into an inline style tag
I want to be sure that layers are defined in the right order, so I moved the
layer definition into an inline style tag that comes before any styles are
loaded. This way I don't have to worry if the browser might move on to parse the
inline pages styles before its finished fetching and parsing the shared
stylesheet.

I also introduced a new layer for page styles that has higher precedence than
component styles, but less precedence than utilities.