Add the webmentions styles to the article page
I'd rather do these as component styles in the webmentions include, but for now
that doesn't seem to be working. I think it might work better with layout
chaining, because I suspect that Nunjucks is following the template inheritance
chain and working its way from the outside in, based on log messages.