Move the portfolio styles into page styles for the layout
There's an error somewhere in here about an empty selector, but the error
reporting from LightningCSS is pretty poor at the moment, so I'm not quite sure
what the issue is. I'm moving it into page styles anyway and just commenting it
out so I can clear out the last of the Sass files. I'm thinking I may move my
portfolio and resume to a separate website anyway, so they're currently not
linked from anywhere anyway.