Fix wrapping with arrow on narrow viewports
Switch to using a hidden span separated from the link by a non-breaking
zero-width space to ensure that the arrow doesn't end up on a line of its own.
When it does, it makes the whitespace look off when it's invisible, and looks
weird when it appears by itself on hover. This forces it to wrap with the last
word so that neither of those things happens.