The arrow has to be inline to prevent breaks
The non-breaking characters don't work with inline-block elements, so I have to
keep the span as an inline. Which means I can't use a transform to move it, so
I'm using position relative and animating the left property. I don't think this
is going to cause any performance issues; it's a small animation, and I don't
think animating left causes a reflow for position relative anyway.