Hide the labels by default in the theme toggle
Since Firefox 110 is supposed to be out tomorrow with container query support, I
think I'll progressively enhance this component with container queries. So by
default it will be an icon-only display with the accessible names hidden from
view. If container queries are supported and there's enough space, I'll unhide
the labels.