Rely on theme.js for managing the theme
Instead of toggling the theme from within the web component, use the functions
from theme.js. This way, we can load the theme.js code synchronously (should
probably inline it) and set the visitor's preferred theme before we have a flash
of another theme. And we don't duplicate code that manages the themes.