Convert webmentions partial to a WebC component
I’m actually leaving a bit of the Nunjucks in place inside this component using
an 11ty:type=njk template because of the loop here. The reason to convert this
to a WebC component, though, is to eliminate the homegrown {% style %} shortcode
I was using to inline page styles. With a WebC component, I can just use
webc:bucket in a <style> tag and it’s all taken care of. This means I can
eliminate a good bit of code from the site.

I am dropping the blobs for now and resorting to a boring circle for the
webmention avatars, because I’m too tired to figure out how to incorporate that
in a sensible way, and because it doesn’t work all that well in dark mode
anyway.