Set min-block-size to fill the viewport
Instead of setting height to 100% on both the body and the root element, set
min-block-size on just the body. Prefer the newer dvb unit to account for
disappearing location bars, with a fallback to 100%. Min size so we don’t end up
with weird scroll behavior, we just want to be able to pull the footer down when
the viewport is very tall.