Quick fix to images in figures
The horizontal alignment of the images in the Taipei Zoo post was off because
the image tags need margin-inline: auto. This is totally unrelated to the rest
of the changes in this branch, but I want it fixed, I’m tired, and I don’t want
to go through another build.