Style the skip link a little
I darkened the text a little, but it’s not that noticeable because I want to
maintain a AAA contrast level. I also added the underline back and aligned it to
the bottom of the grid on wide monitors, which looks a little better, because of
how the theme toggle is offset.