Remove the heading from the plain layout
Plain layouts have no heading, so I just put an h1 in the markdown. These pages
are more likely to use an h1 that is different from the page title, so this just
makes life easier. On the off chance that the page needs the same title, I can
just write it that way.