Convert the base template to Liquid
I am going to keep the theme toggle as a WebC component for now because I
think it actually makes more sense that way. There’s a gotcha with putting WebC
components with asset bundling inside of a Liquid partial, because Liquid
doesn't get the parent context, so you have to pass in the `page` variable for
the bucketing to work.