Convert main layout to a container
Instead of setting max-inline-size on everything as a way of creating
breakout elements without using grid or negative margins, just stick
everything inside a container. This way I don't have to remember to add new
elements to selectors that set max-inline-size (such as `dl`) when I introduce
them in a post.

This fixes the alignment issue with the description list in My Productivity
Stream. I think I'll progressively enhance the layout with grid and maybe
subgrid to allow for breakout elements.

I may be able to remove the .prose class now.