Use justfile to subset fonts
Replace the shell script for font subsetting with a justfile. I can move all of
my tasks to the justfile instead of NPM scripts so that they're all in one
place, and that let's me keep all of the site management in one place. I also
added Atkinson Hyperlegible partially as a test for Just, but also because I'm
thinking about replacing Epilogue with it.